Introduce etl::exit()
For CUDA, it's not possible to rely on static destructor since it is
possible that the CUDA driver will be shut down before the static
destructor is called. Therefore, when ETL_GPU_POOL is used, it is
necessary to use etl::exit() or ETL_PROLOGUE in order to clean up
correctly all the resources.